cleanup(gradle): keep every part of dependency resolution in one file
flattenDependsOn/flattenValues and getDependsOnTask move to
SafeTaskResolution.kt, which now owns the mechanism end to end; TaskUtils.kt
only consumes it. The union of resolved tasks with the raw Task objects in
dependsOn is dropped: the safe resolution already yields every one of them,
and its failure path keeps exactly those, so the union and its second cache
duplicated what resolveDependsOn's cache holds.
